Refactor media control and testing functionality
- Updated skip ad functionality in AdsTester.cs to send user actions.
- Added new tests in MediaChannelTester.cs for media control features.
- Introduced queue management tests in QueueTester.cs for various operations.
- Improved user action handling in MediaChannel.cs; marked SendUserActionAsync as obsolete.
- Removed EditTracksInfoRequest and StreamTransferMessage from serialization context in SharpcasteSerializationContext.cs.
- Simplified EditTracksInfoMessage structure for active track IDs and text track styles.
- Refactored UserActionMessage to inherit from a new base class.
- Changed PlaybackRate type from int to double in MediaStatus.cs for precision.
- Made several properties nullable in MusicTrackMetadata.cs for flexible metadata handling.

+++ Sharpcaster/Models/Media/MediaStatus.cs
@@ -20,7 +20,7 @@ namespace Sharpcaster.Models.Media
         /// Gets or sets the playback rate
         /// </summary>
         [JsonPropertyName("playbackRate")]
-        public int PlaybackRate { get; set; }
+        public double PlaybackRate { get; set; }
 
         /// <summary>
         /// Gets or sets the player state
@@ -81,6 +81,7 @@ namespace Sharpcaster.Models.Media
 
         /// <summary>
         /// Gets or sets the queue data
+        /// Only contains two queue items: the current item and the next item
         /// </summary>
         [JsonPropertyName("queueData")]
         public QueueData QueueData { get; set; }
